requirements.txt
streamlit
groq
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-qdrant
qdrant-client
sentence-transformers
transformers
torch
sentencepiece
